artifacts/streamlit-app/requirements.txt
streamlit==1.41.1
pandas==2.2.3
numpy==2.2.1
matplotlib==3.10.0
seaborn==0.13.2
plotly==5.24.1
scipy==1.15.0
openpyxl==3.1.5
python-docx==1.1.2
reportlab==4.2.5
scikit-learn==1.6.0
pillow==11.1.0
xlrd==2.0.1
statsmodels==0.14.4
pyrebase4
sentence-transformers
fastapi==0.110.0
uvicorn==0.23.2
python-multipart==0.0.6
